pyproject.toml
[project]
name = "sekam2"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiodns>=3.5.0",
    "aiohttp>=3.12.15",
    "discord-py>=2.6.3",
    "matplotlib>=3.10.6",
    "mysqlclient>=2.2.7",
    "pillow>=11.3.0",
    "pilmoji>=2.0.4",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = ["ruff>=0.13.3"]

[tool.ruff.lint]
# 追加ルール
extend-select = [
    "UP",     # 古い書き方を指摘
    "RUF013", # 型注釈でNoneが許可されていないのにNoneが入っている箇所を指摘
]

# 無視するルール
ignore = [
    # openの不要なモード指定を指摘("r", "wt" → "w" など)
    # 明示して困ることはないので無視
    "UP015",
]
